Making the rotation arrow (the compass) prettier.
At first we had a simple unicode character.

Then I found a better character, but it was always facing right.

Then I converted that character into an SVG path, 32x32 (actually 30x32).

But upon rotation it looked weird. Why? Because it was unbalanced. The lowest points were further away from the center than the top point. This can be clearly seen by drawing a circle centered in the middle of the arrow. Then I fixed the arrow to fit inside this imaginary circle.

It looks SO good now... :)

Note: Android 4.x WebView rendering engine rotates the arrow off the expected center. It is rendered correctly in Android 5, in Firefox, in Chrome. I can't test on Safari.